mock-consensus: 🌙 block builder provides `execute` (review)
https://github.com/penumbra-zone/penumbra/pull/3822#discussion_r1490286231
